modified: added CRUD for Collection

+++ src/main/java/com/examplenewstack/newstack/book/dto/BookRequestDTO.java
@@ -33,7 +33,7 @@ public record BookRequestDTO(
         @NotNull(message = "A quantidade de livros disponível não pode ser vazia!")
         int disponibility_quantity,
 
-        Long collectionId,
+        int collectionId,
 
         Long employeeId
 ) {
